+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 1.1.0
+
+* Added `HoldTimeoutGestureRecognizer` and `HoldTimeoutDetector`.
+
 ## 1.0.0
 
 * Bumped to stable release.
 
 ## [1.0.0-dev] - 2021-02-23
 
-* Updated to support null-safety
+* Updated to support null-safety.
 
 ## [0.0.3] - 2019-05-14
 
@@ -12,8 +16,8 @@
 
 ## [0.0.2] - 2018-09-13
 
-* Updated README and example to keep in line with Flutter's default app
+* Updated README and example to keep in line with Flutter's default app.
 
 ## [0.0.1] - 2018-09-13
 
-* First version
+* First version.